+++ src/modules/qrCode/qrCode.controller.ts
@@ -6,6 +6,8 @@ import {
   type QrCodeDeleteType,
 } from './qrCode.schema';
 import { type QrCodeService } from './qrCode.service';
+import { qrCodeOwnership } from '../../middleware/ownership';
+import verifyOwnership from '../../middleware/verifyOwnership';
 
 export class QrCodeController {
   constructor(private readonly qrCodeService: QrCodeService) {}
@@ -26,12 +28,22 @@ export class QrCodeController {
     req: Request<unknown, unknown, QrCodeCreateType>,
     res: Response,
   ) {
+    verifyOwnership(
+      await qrCodeOwnership(Number(req.user.id), req.body.shopId),
+      req,
+    );
+
     const newQrCode = await this.qrCodeService.create(req.body);
 
     return res.status(200).json(newQrCode);
   }
 
   async deleteQrCode(req: Request<QrCodeDeleteType>, res: Response) {
+    verifyOwnership(
+      await qrCodeOwnership(Number(req.user.id), req.params.id),
+      req,
+    );
+
     const deletedQrCode = await this.qrCodeService.delete(req.params.id);
 
     return res.status(200).json(deletedQrCode);
